Partially Confirmatory Item Response Model

pcirm is a partially confirmatory approach to item response models (Chen, 2020), which estimates the intercept for continuous and dichotomous data. Similar to PCFA and GPCFA, there are two major model variants with different constraints for identification. One assumes local independence (LI) with a more exploratory tendency, which can be also called the E-step. The other allows local dependence (LD) with a more confirmatory tendency, which can be also called the C-step. Parameters are obtained by sampling from the posterior distributions with the Markov chain Monte Carlo (MCMC) techniques. Different Bayesian Lasso methods are used to regularize the loading pattern and LD. The estimation results can be summarized with summary.lawbl and the factorial eigenvalue can be plotted with plot_eigen.

pcirm(
  dat,
  Q,
  LD = TRUE,
  cati = NULL,
  PPMC = FALSE,
  burn = 5000,
  iter = 5000,
  update = 1000,
  missing = NA,
  rseed = 12345,
  digits = 4,
  alas = FALSE,
  verbose = FALSE
)

Arguments

dat
A \(N \times J\) data matrix or data.frame consisting of the responses of \(N\) individuals to \(J\) items. Only continuous data are supported currently.

Q
A \(J \times K\) design matrix for the loading pattern with \(K\) factors and \(J\) items. Elements are 1, -1, and 0 for specified, unspecified, and zero-fixed loadings, respectively. For models with LI or the E-step, one can specify a few (e.g., 2) loadings per factor. For models with LD or the C-step, the sufficient condition of one specified loading per item is suggested, although there can be a few items without any specified loading (Chen, Guo, Zhang, & Pan, 2020). See Examples.

LD
logical; TRUE for allowing LD (model with LD or C-step).

cati
The set of dichotomous items in sequence number (i.e., 1 to \(J\)); NULL for no and -1 for all items (default is NULL).

PPMC
logical; TRUE for conducting posterior predictive model checking.

burn
Number of burn-in iterations before posterior sampling.

iter
Number of formal iterations for posterior sampling (> 0).

update
Number of iterations to update the sampling information.

missing
Value for missing data (default is NA).

rseed
An integer for the random seed.

digits
Number of significant digits to print when printing numeric values.

alas
logical; for adaptive Lasso or not. The default is FALSE, which seems slightly stabler.

verbose
logical; to display the sampling information every update or not.

Feigen: Eigenvalue for each factor.

NLA_lg3: Number of Loading estimates > .3 for each factor.

Shrink: Ave. shrinkage parameter (for adaptive LASSO) for each factor.

Adj PSR: Adjusted PSR for each factor.

Ave. Int.: Ave. item intercept.

LD>.2 >.1 LD>.2 >.1: # of LD terms larger than .2 and .1, and LD's shrinkage parameter.

Value

pcirm returns an object of class lawbl with item intercepts. It contains a lot of information about the posteriors that can be summarized using summary.lawbl. The factorial eigenvalue can be plotted with plot_eigen.

References

Chen, J. (2020). A partially confirmatory approach to the multidimensional item response theory with the Bayesian Lasso. Psychometrika. 85(3), 738-774. DOI:10.1007/s11336-020-09724-3.

Examples

# \donttest{
####################################
#  Example 1: Estimation with LI   #
####################################

dat <- sim18cfa1$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1
Q

#>       [,1] [,2] [,3]
#>  [1,]    1   -1   -1
#>  [2,]    1   -1   -1
#>  [3,]   -1   -1   -1
#>  [4,]   -1   -1   -1
#>  [5,]   -1   -1   -1
#>  [6,]   -1   -1   -1
#>  [7,]   -1   -1   -1
#>  [8,]   -1   -1   -1
#>  [9,]   -1    1   -1
#> [10,]   -1    1   -1
#> [11,]   -1   -1   -1
#> [12,]   -1   -1   -1
#> [13,]   -1   -1    1
#> [14,]   -1   -1    1
#> [15,]   -1   -1   -1
#> [16,]   -1   -1   -1
#> [17,]   -1   -1   -1
#> [18,]   -1   -1   -1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 5000
#> 
#> $Iteration
#> [1] 5000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.020896   1.095994
#> F2   1.366068   2.207132
#> F3   1.033023   1.091168
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd      lower     upper sig
#> var1     1 1 0.6823035 0.07073808 0.54899356 0.8323660   1
#> var2     2 1 0.6859118 0.07071041 0.55228126 0.8318734   1
#> var3     3 1 0.9023841 0.08948767 0.70872624 1.0718473   1
#> var4     4 1 0.8751598 0.08628161 0.69100012 1.0388933   1
#> var5     5 1 0.7111464 0.07351279 0.58054074 0.8612567   1
#> var6     6 1 0.7306655 0.07437293 0.59033593 0.8777524   1
#> var17   17 1 0.3213983 0.09826327 0.11600496 0.5190424   1
#> var18   18 1 0.3414977 0.09756148 0.13037509 0.5295699   1
#> var23    5 2 0.2445881 0.10044092 0.03499408 0.4348215   1
#> var24    6 2 0.2224391 0.10183027 0.01509840 0.4196578   1
#> var25    7 2 0.6097788 0.07538802 0.47271494 0.7657608   1
#> var26    8 2 0.5992284 0.07160314 0.46738557 0.7453782   1
#> var27    9 2 0.8713781 0.09602151 0.71155996 1.0685804   1
#> var28   10 2 0.9102360 0.09605361 0.74387818 1.1109094   1
#> var29   11 2 0.6457283 0.07706906 0.49406204 0.7917918   1
#> var30   12 2 0.6583718 0.07867359 0.51522931 0.8200805   1
#> var47   11 3 0.2791487 0.10187503 0.07171180 0.4531849   1
#> var48   12 3 0.2938125 0.10352699 0.08590065 0.4749748   1
#> var49   13 3 0.6760728 0.07397181 0.53439753 0.8256898   1
#> var50   14 3 0.5923485 0.06880536 0.45946588 0.7276771   1
#> var51   15 3 0.8719171 0.08435527 0.71865693 1.0527674   1
#> var52   16 3 0.8939878 0.08592217 0.74112391 1.0799171   1
#> var53   17 3 0.6446634 0.07617622 0.48698461 0.7792120   1
#> var54   18 3 0.6339845 0.07695066 0.48351238 0.7790726   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.6823035 0.0000000 0.0000000
#> I2  0.6859118 0.0000000 0.0000000
#> I3  0.9023841 0.0000000 0.0000000
#> I4  0.8751598 0.0000000 0.0000000
#> I5  0.7111464 0.2445881 0.0000000
#> I6  0.7306655 0.2224391 0.0000000
#> I7  0.0000000 0.6097788 0.0000000
#> I8  0.0000000 0.5992284 0.0000000
#> I9  0.0000000 0.8713781 0.0000000
#> I10 0.0000000 0.9102360 0.0000000
#> I11 0.0000000 0.6457283 0.2791487
#> I12 0.0000000 0.6583718 0.2938125
#> I13 0.0000000 0.0000000 0.6760728
#> I14 0.0000000 0.0000000 0.5923485
#> I15 0.0000000 0.0000000 0.8719171
#> I16 0.0000000 0.0000000 0.8939878
#> I17 0.3213983 0.0000000 0.6446634
#> I18 0.3414977 0.0000000 0.6339845

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

####################################
#  Example 2: Estimation with LD   #
####################################

Q<-matrix(-1,J,K);
Q[1:6,1]<-Q[7:12,2]<-Q[13:18,3]<-1
Q

#>       [,1] [,2] [,3]
#>  [1,]    1   -1   -1
#>  [2,]    1   -1   -1
#>  [3,]    1   -1   -1
#>  [4,]    1   -1   -1
#>  [5,]    1   -1   -1
#>  [6,]    1   -1   -1
#>  [7,]   -1    1   -1
#>  [8,]   -1    1   -1
#>  [9,]   -1    1   -1
#> [10,]   -1    1   -1
#> [11,]   -1    1   -1
#> [12,]   -1    1   -1
#> [13,]   -1   -1    1
#> [14,]   -1   -1    1
#> [15,]   -1   -1    1
#> [16,]   -1   -1    1
#> [17,]   -1   -1    1
#> [18,]   -1   -1    1

m1 <- pcfa(dat = dat, Q = Q)
summary(m1) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] TRUE
#> 
#> $`Burn in`
#> [1] 5000
#> 
#> $Iteration
#> [1] 5000
#> 
#> $`No. of sig lambda`
#> [1] 21
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.022491   1.024425
#> F2   1.052765   1.195116
#> F3   1.003639   1.018012
#> 
#> $`No. of sig LD terms`
#> [1] 6
#> 

summary(m1, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7080049 0.0000000 0.0000000
#> I2  0.7156364 0.0000000 0.0000000
#> I3  0.7456512 0.0000000 0.0000000
#> I4  0.7194150 0.0000000 0.0000000
#> I5  0.7631656 0.0000000 0.0000000
#> I6  0.7470119 0.2130088 0.0000000
#> I7  0.0000000 0.7466830 0.0000000
#> I8  0.0000000 0.6387746 0.0000000
#> I9  0.0000000 0.7029578 0.0000000
#> I10 0.0000000 0.7463125 0.0000000
#> I11 0.0000000 0.7627909 0.0000000
#> I12 0.0000000 0.7680904 0.0000000
#> I13 0.0000000 0.0000000 0.7491191
#> I14 0.0000000 0.0000000 0.6881314
#> I15 0.0000000 0.0000000 0.7134234
#> I16 0.0000000 0.0000000 0.7443122
#> I17 0.2613409 0.0000000 0.7073008
#> I18 0.2620075 0.0000000 0.6907499

summary(m1, what = 'offpsx') #summarize significant LD terms

#>        row col       est         sd     lower     upper sig
#> var14   14   1 0.2755861 0.03573670 0.2036052 0.3432464   1
#> var24    7   2 0.2727970 0.03925801 0.1976261 0.3516715   1
#> var37    4   3 0.2969177 0.04441451 0.2058291 0.3846609   1
#> var111  13   8 0.2901256 0.04131129 0.2023568 0.3640477   1
#> var118  10   9 0.3023453 0.05692644 0.1876921 0.4080570   1
#> var163  16  15 0.2780553 0.06073809 0.1575662 0.3954573   1

# }