Partially Exploratory Factor Analysis

PEFA is a partially exploratory approach to factor analysis, which can incorporate partial knowledge together with unknown number of factors, using bi-level Bayesian regularization. When partial knowledge is not needed, it reduces to the fully exploratory factor analysis (FEFA; Chen, 2021). A large number of factors can be imposed for selection where true factors will be identified against spurious factors. The loading vector is reparameterized to tackle model sparsity at the factor and loading levels with the multivariate spike and slab priors. Parameters are obtained by sampling from the posterior distributions with the Markov chain Monte Carlo (MCMC) techniques. The estimation results can be summarized with summary.lawbl and the trace or density of the posterior can be plotted with plot_lawbl.

pefa(
  dat,
  Q = NULL,
  K = 8,
  mjf = 3,
  PPMC = FALSE,
  burn = 5000,
  iter = 5000,
  update = 1000,
  rseed = 12345,
  digits = 4,
  alas = FALSE,
  verbose = FALSE
)

Arguments

dat
A \(N \times J\) data matrix or data.frame consisting of the responses of \(N\) individuals to \(J\) items.

Q
A \(J \times K\) design matrix for the loading pattern with \(K\) factors and \(J\) items for PEFA. Elements are 1, -1, and 0 for specified, unspecified, and zero-fixed loadings, respectively. It's not needed for FEFA, which is the default. See Examples.

K
Maximum number of factors for selection under FEFA. Not used for PEFA.

mjf
Minimum number of items per factor.

PPMC
logical; TRUE for conducting posterior predictive model checking.

burn
Number of burn-in iterations before posterior sampling.

iter
Number of formal iterations for posterior sampling (> 0).

update
Number of iterations to update the sampling information.

rseed
An integer for the random seed.

digits
Number of significant digits to print when printing numeric values.

verbose
logical; to display the sampling information every update or not.

Feigen: Eigenvalue for each factor.

NLA_le3: Number of Loading estimates >= .3 for each factor.

Shrink: Shrinkage parameter for each factor.

True Fa: Is the factor identified as true or not.

Adj PSR: Adjusted PSR for each factor.

Value

pcfa returns an object of class lawbl without item intercepts. It contains a lot of information about the posteriors that can be summarized using summary.lawbl.

References

Chen, J. (2021). A Bayesian regularized approach to exploratory factor analysis in one step. Structural Equation Modeling: A Multidisciplinary Journal. DOI: 10.1080/10705511.2020.1854763.

Examples

# \donttest{
#####################################################
#  Example 1: Fully EFA                             #
#####################################################

dat <- sim18cfa0$dat

m0 <- pefa(dat = dat, K=5, burn = 2000, iter = 2000,verbose = TRUE)

#> 
#> Tot. Iter = 1000
#>           [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  3.3600 3.1506 3.0205 0.000e+00 0.000e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.2665 0.2595 0.1969 9.241e-84 1.348e-83
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 2000
#>           [,1]   [,2]   [,3]     [,4]    [,5]
#> Feigen  3.3740 3.2687 3.0413 0.00e+00 0.0e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.00e+00 0.0e+00
#> iShrink 0.2236 0.2703 0.1598 1.24e-84 4.1e-84
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 3000
#>           [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  3.2468 3.2243 3.0053 0.000e+00 0.000e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.2113 0.2923 0.1628 6.599e-84 2.866e-84
#> True Fa 1 1 1 0 0
#> Adj PSR 1.038 1.071 1.003
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 4000
#>          [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  3.237 3.2180 3.1719 0.000e+00 0.000e+00
#> NLA_le3 8.000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.195 0.2945 0.2037 5.518e-84 5.675e-84
#> True Fa 1 1 1 0 0
#> Adj PSR 1.022 1.048 1.016
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#>    user  system elapsed 
#>   17.44    0.00   17.44 

summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 5
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`True Factor`
#> [1]  TRUE  TRUE  TRUE FALSE FALSE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.022412   1.103702
#> F2   1.047502   1.201463
#> F3   1.015633   1.016419
#> 

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3] [,4] [,5]
#> I1  0.0000000 0.0000000 0.6667297    0    0
#> I2  0.0000000 0.0000000 0.6787770    0    0
#> I3  0.0000000 0.0000000 0.6827262    0    0
#> I4  0.0000000 0.0000000 0.6644396    0    0
#> I5  0.4212348 0.0000000 0.6910990    0    0
#> I6  0.4159950 0.0000000 0.6890032    0    0
#> I7  0.6967177 0.0000000 0.0000000    0    0
#> I8  0.6801511 0.0000000 0.0000000    0    0
#> I9  0.6998491 0.0000000 0.0000000    0    0
#> I10 0.6965932 0.0000000 0.0000000    0    0
#> I11 0.7000259 0.4121509 0.0000000    0    0
#> I12 0.6970445 0.4236665 0.0000000    0    0
#> I13 0.0000000 0.6979444 0.0000000    0    0
#> I14 0.0000000 0.6718842 0.0000000    0    0
#> I15 0.0000000 0.6867237 0.0000000    0    0
#> I16 0.0000000 0.6868011 0.0000000    0    0
#> I17 0.0000000 0.7078016 0.3930731    0    0
#> I18 0.0000000 0.7213827 0.3921211    0    0

summary(m0, what = 'phi') #summarize factorial correlations

#>      row col       est         sd     lower     upper sig
#> var1   2   1 0.2579280 0.04058422 0.1758287 0.3328396   1
#> var2   3   1 0.2502747 0.03600974 0.1761319 0.3179651   1
#> var5   3   2 0.2459183 0.03923047 0.1681724 0.3211863   1

summary(m0, what = 'eigen') #summarize factorial eigenvalue

#>         est         sd    lower    upper sig
#> F1 3.253181 0.11408021 3.024826 3.471741   1
#> F2 3.266066 0.10906184 3.064555 3.484213   1
#> F3 3.075459 0.09858594 2.878956 3.259401   1

##########################################################
#  Example 2: PEFA with two factors partially specified  #
##########################################################

J <- ncol(dat)
K <- 5
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[7:8,2]<-1
Q

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1   -1   -1   -1   -1
#>  [2,]    1   -1   -1   -1   -1
#>  [3,]   -1   -1   -1   -1   -1
#>  [4,]   -1   -1   -1   -1   -1
#>  [5,]   -1   -1   -1   -1   -1
#>  [6,]   -1   -1   -1   -1   -1
#>  [7,]   -1    1   -1   -1   -1
#>  [8,]   -1    1   -1   -1   -1
#>  [9,]   -1   -1   -1   -1   -1
#> [10,]   -1   -1   -1   -1   -1
#> [11,]   -1   -1   -1   -1   -1
#> [12,]   -1   -1   -1   -1   -1
#> [13,]   -1   -1   -1   -1   -1
#> [14,]   -1   -1   -1   -1   -1
#> [15,]   -1   -1   -1   -1   -1
#> [16,]   -1   -1   -1   -1   -1
#> [17,]   -1   -1   -1   -1   -1
#> [18,]   -1   -1   -1   -1   -1

m1 <- pefa(dat = dat, Q = Q,burn = 2000, iter = 2000,verbose = TRUE)

#> 
#> Tot. Iter = 1000
#>           [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  3.2888 3.3304 3.2088 0.000e+00 0.000e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.5268 0.5268 0.4651 2.769e-84 3.632e-84
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 2000
#>           [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  3.1966 3.4909 3.1559 0.000e+00 0.000e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.4174 0.4174 0.3526 1.601e-84 2.383e-84
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 3000
#>           [,1]   [,2]   [,3]      [,4]      [,5]
#> Feigen  2.9410 3.5358 3.1515 0.000e+00 0.000e+00
#> NLA_le3 8.0000 8.0000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.6008 0.6008 0.2938 2.119e-84 5.767e-84
#> True Fa 1 1 1 0 0
#> Adj PSR 1.02 1.008 1.001
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#> 
#> Tot. Iter = 4000
#>          [,1]  [,2]   [,3]      [,4]      [,5]
#> Feigen  3.195 3.284 3.2894 0.000e+00 0.000e+00
#> NLA_le3 8.000 8.000 8.0000 0.000e+00 0.000e+00
#> iShrink 0.615 0.615 0.2264 4.429e-84 1.104e-83
#> True Fa 1 1 1 0 0
#> Adj PSR 1.001 1.007 1.086
#> [1] "ROW: LA overflow,sign switch,MJF total, current, reset"
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    0    0    0    0
#> [2,]    0    0    0    0    0
#> [3,]    0    0    0    0    0
#> [4,]    0    0    0    0    0
#> [5,]    0    0    0    0    0
#>    user  system elapsed 
#>   17.41    0.00   17.41 

summary(m1)

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 5
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`True Factor`
#> [1]  TRUE  TRUE  TRUE FALSE FALSE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.000528   1.000553
#> F2   1.006989   1.027575
#> F3   1.086365   1.303004
#> 

summary(m1, what = 'qlambda')

#>          [,1]      [,2]      [,3] [,4] [,5]
#> I1  0.6735315 0.0000000 0.0000000    0    0
#> I2  0.6843638 0.0000000 0.0000000    0    0
#> I3  0.6892692 0.0000000 0.0000000    0    0
#> I4  0.6707357 0.0000000 0.0000000    0    0
#> I5  0.6950772 0.4223813 0.0000000    0    0
#> I6  0.6948685 0.4182749 0.0000000    0    0
#> I7  0.0000000 0.7186032 0.0000000    0    0
#> I8  0.0000000 0.6919956 0.0000000    0    0
#> I9  0.0000000 0.7087884 0.0000000    0    0
#> I10 0.0000000 0.7039691 0.0000000    0    0
#> I11 0.0000000 0.7072619 0.4077364    0    0
#> I12 0.0000000 0.7056352 0.4193172    0    0
#> I13 0.0000000 0.0000000 0.6992504    0    0
#> I14 0.0000000 0.0000000 0.6725351    0    0
#> I15 0.0000000 0.0000000 0.6873442    0    0
#> I16 0.0000000 0.0000000 0.6870559    0    0
#> I17 0.3958852 0.0000000 0.7080387    0    0
#> I18 0.3949583 0.0000000 0.7221333    0    0

summary(m1, what = 'phi')

#>      row col       est         sd     lower     upper sig
#> var1   2   1 0.2557402 0.03488180 0.1817713 0.3186925   1
#> var2   3   1 0.2448117 0.03720091 0.1758008 0.3178213   1
#> var5   3   2 0.2634231 0.03702354 0.1901555 0.3340776   1

summary(m1,what='eigen')

#>         est        sd    lower    upper sig
#> F1 3.128837 0.1218312 2.883722 3.356353   1
#> F2 3.348771 0.1262370 3.101925 3.599454   1
#> F3 3.261802 0.1139679 3.049078 3.473366   1

# }