Simulating data with Latent Variable Modeling

sim_lvm can simulate data based on factor analysis or item response models with different response formats (continuous or categorical), loading patterns and residual covariance (local dependence) structures.

sim_lvm(
  N = 1000,
  K = 3,
  ipf = 8,
  cpf = 2,
  lam = 0.7,
  lac = 0.3,
  phi = 0.5,
  ph1 = -1,
  ecr = 0,
  ome_out = FALSE,
  cati = NULL,
  noc = c(4),
  misp = 0,
  rseed = 333,
  necw = K,
  necb = K,
  add_ind = c(),
  add_la = 0.5,
  add_phi = 0,
  zero_it = 0
)

Arguments

N
Sample size.

K
Number of factors.

ipf
Items per factor.

cpf
Cross-loadings per factor.

lam
Number of formal iterations for posterior sampling.

lac
Number of iterations to update the sampling information.

phi
Homogeneous correlations between any two factors.

ph1
Correlation between factor 1 and 2 (if it's different from phi.

ecr
Residual correlation (local dependence).

ome_out
Output factor score or not.

cati
The set of categorical (polytomous) items in sequence number (i.e., 1 to \(J\)); NULL for no and -1 for all (default is NULL).

noc
Number of categories for categorical items

misp
Proportion of missingness.

rseed
An integer for the random seed.

necw
Number of within-factor local dependence.

necb
Number of between-factor local dependence.

add_ind
(Additional) minor factor with cross-loadings.

add_la
Value of cross-loadings on (Additional) minor factor.

add_phi
Correlations between (Additional) minor factor and other factors.

zero_it
Surplus items with zero loading.

Value

An object of class list containing the data, loading, and factorial correlation matrix.

Examples

# for continuous data with cross-loadings and local dependence effect .3
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.3,ecr=.3,rseed = 123)
head(out$dat)

#>             [,1]       [,2]       [,3]        [,4]       [,5]         [,6]
#> [1,] -0.36689395 -0.4881216 -1.7566272 -0.50482295 -0.4684937 -1.297411644
#> [2,]  1.29615435  1.1647742  0.3647204 -0.03480086  1.3327210  0.849699024
#> [3,] -1.88314706  0.1651701 -0.4958659 -0.86045126 -0.6903347  0.338318624
#> [4,] -0.05165739  0.8702513  0.2634152 -0.37815424  0.2265568  0.888059924
#> [5,]  0.11355713 -1.8976949 -2.5072503 -0.73721254  1.0193241 -0.001280834
#> [6,] -2.01181093 -0.7932417  0.1727423 -0.29495085 -1.3349318 -0.338248749
#>            [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#> [1,] -0.5173644 -0.4091269 -1.7824866 -0.2460811 -1.0117432 -0.8083742
#> [2,]  1.5628364  2.4832469  2.2553101  2.1472036  1.4055864  1.7997063
#> [3,]  2.1703316  1.0343184  0.1287202  0.7294599  0.7892556  0.7092318
#> [4,]  1.1868684  0.4187861  2.2694918  1.9012874  1.9883522  1.6677417
#> [5,] -0.4761872 -0.4433155  1.4434187  1.5892937  1.3876958  1.3504176
#> [6,]  0.7946051  1.9575439  0.5235337  1.2571445  1.2672767  0.8616947
#>            [,13]      [,14]       [,15]      [,16]        [,17]       [,18]
#> [1,]  0.67924455  1.3030044  0.91846529 -0.1301423 -0.139432271  0.43771563
#> [2,]  0.26520885  0.2609328  0.46525357  0.7109910  0.002888877  1.15870865
#> [3,]  1.03667459 -0.7095409 -0.90312500  0.3296742  0.077149576 -0.77161803
#> [4,]  0.02234946  0.9213876  1.51360868  1.9036678  0.746768729  1.14153154
#> [5,] -0.80397103  0.1202811 -0.05229181  1.0703991  0.247518511  0.49762899
#> [6,]  1.42219402 -0.9986785 -0.44821014 -0.5433925  0.302156652  0.02627775

out$MLA

#>       [,1] [,2] [,3]
#>  [1,]  0.7  0.0  0.0
#>  [2,]  0.7  0.0  0.0
#>  [3,]  0.7  0.0  0.0
#>  [4,]  0.7  0.0  0.0
#>  [5,]  0.7  0.3  0.0
#>  [6,]  0.7  0.3  0.0
#>  [7,]  0.0  0.7  0.0
#>  [8,]  0.0  0.7  0.0
#>  [9,]  0.0  0.7  0.0
#> [10,]  0.0  0.7  0.0
#> [11,]  0.0  0.7  0.3
#> [12,]  0.0  0.7  0.3
#> [13,]  0.0  0.0  0.7
#> [14,]  0.0  0.0  0.7
#> [15,]  0.0  0.0  0.7
#> [16,]  0.0  0.0  0.7
#> [17,]  0.3  0.0  0.7
#> [18,]  0.3  0.0  0.7

out$ofd_ind

#>      row col
#> [1,]  14   1
#> [2,]   7   2
#> [3,]   4   3
#> [4,]  13   8
#> [5,]  10   9
#> [6,]  16  15

# for categorical data with cross-loadings .4
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.4,cati=-1,noc=4,rseed = 123)
head(out$dat)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    2    2    1    2    2    2    2    3    3     2     2     2     3     3
#> [2,]    3    3    3    4    4    3    3    3    3     4     4     4     3     3
#> [3,]    2    3    3    2    2    3    3    4    3     2     3     3     2     3
#> [4,]    2    3    3    2    3    3    3    3    3     3     4     4     4     3
#> [5,]    1    3    3    3    3    3    3    3    4     3     3     3     3     2
#> [6,]    3    3    2    2    2    2    3    3    3     3     3     3     3     3
#>      [,15] [,16] [,17] [,18]
#> [1,]     3     3     3     3
#> [2,]     3     3     3     3
#> [3,]     2     2     2     2
#> [4,]     3     4     3     3
#> [5,]     4     3     3     3
#> [6,]     2     2     2     2

out$MLA

#>       [,1] [,2] [,3]
#>  [1,]  0.7  0.0  0.0
#>  [2,]  0.7  0.0  0.0
#>  [3,]  0.7  0.0  0.0
#>  [4,]  0.7  0.0  0.0
#>  [5,]  0.7  0.4  0.0
#>  [6,]  0.7  0.4  0.0
#>  [7,]  0.0  0.7  0.0
#>  [8,]  0.0  0.7  0.0
#>  [9,]  0.0  0.7  0.0
#> [10,]  0.0  0.7  0.0
#> [11,]  0.0  0.7  0.4
#> [12,]  0.0  0.7  0.4
#> [13,]  0.0  0.0  0.7
#> [14,]  0.0  0.0  0.7
#> [15,]  0.0  0.0  0.7
#> [16,]  0.0  0.0  0.7
#> [17,]  0.4  0.0  0.7
#> [18,]  0.4  0.0  0.7

out$ofd_ind

#>      row col

# \donttest{
dat <- out$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1
Q

#>       [,1] [,2] [,3]
#>  [1,]    1   -1   -1
#>  [2,]    1   -1   -1
#>  [3,]   -1   -1   -1
#>  [4,]   -1   -1   -1
#>  [5,]   -1   -1   -1
#>  [6,]   -1   -1   -1
#>  [7,]   -1   -1   -1
#>  [8,]   -1   -1   -1
#>  [9,]   -1    1   -1
#> [10,]   -1    1   -1
#> [11,]   -1   -1   -1
#> [12,]   -1   -1   -1
#> [13,]   -1   -1    1
#> [14,]   -1   -1    1
#> [15,]   -1   -1   -1
#> [16,]   -1   -1   -1
#> [17,]   -1   -1   -1
#> [18,]   -1   -1   -1

m0 <- pcfa(dat = dat, cati = -1, Q = Q, LD = FALSE,burn = 3000, iter = 3000,verbose=TRUE)

#> 
#> Tot. Iter = 1000
#>          [,1]   [,2]  [,3]
#> Feigen  3.007  4.402 5.418
#> NLA_lg3 6.000 10.000 8.000
#> Shrink  2.609  2.609 2.609
#> Ave. Thd: -1e+200 -1.867 -0.2897 1.214 1e+200
#> Acc Rate: 0.3601
#> 
#> Tot. Iter = 2000
#>          [,1]  [,2]  [,3]
#> Feigen  3.629 3.216 3.532
#> NLA_lg3 8.000 8.000 8.000
#> Shrink  2.518 2.518 2.518
#> Ave. Thd: -1e+200 -1.809 -0.2319 1.225 1e+200
#> Acc Rate: 0.3551
#> 
#> Tot. Iter = 3000
#>          [,1]  [,2]  [,3]
#> Feigen  3.923 3.162 3.956
#> NLA_lg3 8.000 6.000 8.000
#> Shrink  2.889 2.889 2.889
#> Ave. Thd: -1e+200 -1.748 -0.2131 1.287 1e+200
#> Acc Rate: 0.3536
#> 
#> Tot. Iter = 4000
#>          [,1]  [,2]  [,3]
#> Feigen  3.828 3.263 3.530
#> NLA_lg3 8.000 8.000 8.000
#> Shrink  3.719 3.719 3.719
#> Adj PSR 1.07 1.01 2.087
#> Ave. Thd: -1e+200 -1.691 -0.1414 1.325 1e+200
#> Acc Rate: 0.3523
#> 
#> Tot. Iter = 5000
#>          [,1]  [,2]  [,3]
#> Feigen  3.372 3.501 3.618
#> NLA_lg3 8.000 8.000 8.000
#> Shrink  2.891 2.891 2.891
#> Adj PSR 1.649 1.075 1.019
#> Ave. Thd: -1e+200 -1.693 -0.1558 1.344 1e+200
#> Acc Rate: 0.3531
#> 
#> Tot. Iter = 6000
#>          [,1]  [,2]  [,3]
#> Feigen  3.580 3.049 3.755
#> NLA_lg3 8.000 8.000 8.000
#> Shrink  3.445 3.445 3.445
#> Adj PSR 1.166 1.076 1.013
#> Ave. Thd: -1e+200 -1.706 -0.1385 1.371 1e+200
#> Acc Rate: 0.3546
#> 
#> #Sign change: 0 0 0 0 0 0
#>    user  system elapsed 
#>   100.5     0.0   100.5 

summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 3000
#> 
#> $Iteration
#> [1] 3000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.166455   1.579199
#> F2   1.075824   1.158405
#> F3   1.012916   1.019371
#> 
#> $`Cat Items`
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18
#> 
#> $`max No. of categories`
#> [1] 4
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd     lower     upper sig
#> var1     1 1 0.7513928 0.06086839 0.6334580 0.8655604   1
#> var2     2 1 0.7846111 0.05686583 0.6698745 0.8898211   1
#> var3     3 1 0.6963859 0.05187712 0.5934600 0.7976899   1
#> var4     4 1 0.7097179 0.05552289 0.6011540 0.8167845   1
#> var5     5 1 0.6775012 0.05660098 0.5628928 0.7973446   1
#> var6     6 1 0.7027709 0.05506972 0.6002609 0.8216999   1
#> var17   17 1 0.3704505 0.06800001 0.2267468 0.4897480   1
#> var18   18 1 0.3920823 0.07137759 0.2518373 0.5249785   1
#> var23    5 2 0.4521916 0.07264070 0.3108411 0.5951826   1
#> var24    6 2 0.4209032 0.07264477 0.2797739 0.5575893   1
#> var25    7 2 0.7010917 0.05625919 0.5864559 0.8079367   1
#> var26    8 2 0.7058866 0.06302445 0.5917430 0.8368060   1
#> var27    9 2 0.7244635 0.06384414 0.5985317 0.8475623   1
#> var28   10 2 0.7333519 0.05781271 0.6224628 0.8519866   1
#> var29   11 2 0.7221971 0.04970786 0.6329132 0.8231842   1
#> var30   12 2 0.6651844 0.04695533 0.5762356 0.7618485   1
#> var47   11 3 0.4088030 0.06948185 0.2759376 0.5428537   1
#> var48   12 3 0.4445436 0.06305243 0.3218806 0.5631624   1
#> var49   13 3 0.7743668 0.05968387 0.6627701 0.8936835   1
#> var50   14 3 0.7477978 0.05630337 0.6387987 0.8556948   1
#> var51   15 3 0.7058896 0.05872753 0.6030618 0.8323233   1
#> var52   16 3 0.7265640 0.06071449 0.6158904 0.8468757   1
#> var53   17 3 0.7278373 0.05517413 0.6275085 0.8351958   1
#> var54   18 3 0.7498162 0.05851032 0.6349423 0.8528206   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7513928 0.0000000 0.0000000
#> I2  0.7846111 0.0000000 0.0000000
#> I3  0.6963859 0.0000000 0.0000000
#> I4  0.7097179 0.0000000 0.0000000
#> I5  0.6775012 0.4521916 0.0000000
#> I6  0.7027709 0.4209032 0.0000000
#> I7  0.0000000 0.7010917 0.0000000
#> I8  0.0000000 0.7058866 0.0000000
#> I9  0.0000000 0.7244635 0.0000000
#> I10 0.0000000 0.7333519 0.0000000
#> I11 0.0000000 0.7221971 0.4088030
#> I12 0.0000000 0.6651844 0.4445436
#> I13 0.0000000 0.0000000 0.7743668
#> I14 0.0000000 0.0000000 0.7477978
#> I15 0.0000000 0.0000000 0.7058896
#> I16 0.0000000 0.0000000 0.7265640
#> I17 0.3704505 0.0000000 0.7278373
#> I18 0.3920823 0.0000000 0.7498162

# }