Summary method for pcfa objectects

Provide basic information for an PCFA object, and summarize various posteriors.

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

thd: Threshold estimates.

int: Intercept estimates (for pcirm only).

shrink: (Ave) shrinkage for each factor's loadings and LD (if LD in pcfa = T).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

# \donttest{
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 22
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.072878   1.201632
#> F2   2.484116   7.247529
#> F3   1.030576   1.131831
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd       lower     upper sig
#> var1     1 1 0.7380810 0.06628118 0.628051125 0.8731824   1
#> var2     2 1 0.7770108 0.07463361 0.654304394 0.9207123   1
#> var3     3 1 0.7538982 0.07103708 0.636941141 0.8961474   1
#> var4     4 1 0.6924366 0.06427519 0.580317325 0.8224172   1
#> var5     5 1 0.7519420 0.06526289 0.636548078 0.8742827   1
#> var6     6 1 0.7467335 0.06530851 0.629215154 0.8669389   1
#> var17   17 1 0.2731530 0.07289758 0.113970562 0.3984249   1
#> var18   18 1 0.2801235 0.07352431 0.126351423 0.4097882   1
#> var23    5 2 0.2289749 0.09316069 0.005585078 0.3702318   1
#> var24    6 2 0.2427417 0.09454963 0.043129190 0.4063930   1
#> var25    7 2 0.7697333 0.08835142 0.614122188 0.9295117   1
#> var26    8 2 0.7689378 0.09071653 0.621257701 0.9484193   1
#> var27    9 2 0.7711148 0.08814999 0.631163178 0.9419877   1
#> var28   10 2 0.6957037 0.07704221 0.550888909 0.8382743   1
#> var29   11 2 0.7367116 0.09195786 0.578555492 0.9289098   1
#> var30   12 2 0.7497902 0.09358578 0.571542096 0.9244531   1
#> var49   13 3 0.7348116 0.05721002 0.621609825 0.8375784   1
#> var50   14 3 0.7647557 0.05930961 0.649196208 0.8755927   1
#> var51   15 3 0.7561532 0.06015950 0.649150458 0.8874273   1
#> var52   16 3 0.7108223 0.06337808 0.588784788 0.8324290   1
#> var53   17 3 0.7047157 0.05364273 0.598700591 0.8016899   1
#> var54   18 3 0.7099413 0.05282302 0.614491281 0.8091791   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7380810 0.0000000 0.0000000
#> I2  0.7770108 0.0000000 0.0000000
#> I3  0.7538982 0.0000000 0.0000000
#> I4  0.6924366 0.0000000 0.0000000
#> I5  0.7519420 0.2289749 0.0000000
#> I6  0.7467335 0.2427417 0.0000000
#> I7  0.0000000 0.7697333 0.0000000
#> I8  0.0000000 0.7689378 0.0000000
#> I9  0.0000000 0.7711148 0.0000000
#> I10 0.0000000 0.6957037 0.0000000
#> I11 0.0000000 0.7367116 0.0000000
#> I12 0.0000000 0.7497902 0.0000000
#> I13 0.0000000 0.0000000 0.7348116
#> I14 0.0000000 0.0000000 0.7647557
#> I15 0.0000000 0.0000000 0.7561532
#> I16 0.0000000 0.0000000 0.7108223
#> I17 0.2731530 0.0000000 0.7047157
#> I18 0.2801235 0.0000000 0.7099413

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

summary(m0, what = 'all') #summarize all information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 22
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.072878   1.201632
#> F2   2.484116   7.247529
#> F3   1.030576   1.131831
#> 
#> $lambda
#>       Item F       est         sd       lower     upper sig
#> var1     1 1 0.7380810 0.06628118 0.628051125 0.8731824   1
#> var2     2 1 0.7770108 0.07463361 0.654304394 0.9207123   1
#> var3     3 1 0.7538982 0.07103708 0.636941141 0.8961474   1
#> var4     4 1 0.6924366 0.06427519 0.580317325 0.8224172   1
#> var5     5 1 0.7519420 0.06526289 0.636548078 0.8742827   1
#> var6     6 1 0.7467335 0.06530851 0.629215154 0.8669389   1
#> var17   17 1 0.2731530 0.07289758 0.113970562 0.3984249   1
#> var18   18 1 0.2801235 0.07352431 0.126351423 0.4097882   1
#> var23    5 2 0.2289749 0.09316069 0.005585078 0.3702318   1
#> var24    6 2 0.2427417 0.09454963 0.043129190 0.4063930   1
#> var25    7 2 0.7697333 0.08835142 0.614122188 0.9295117   1
#> var26    8 2 0.7689378 0.09071653 0.621257701 0.9484193   1
#> var27    9 2 0.7711148 0.08814999 0.631163178 0.9419877   1
#> var28   10 2 0.6957037 0.07704221 0.550888909 0.8382743   1
#> var29   11 2 0.7367116 0.09195786 0.578555492 0.9289098   1
#> var30   12 2 0.7497902 0.09358578 0.571542096 0.9244531   1
#> var49   13 3 0.7348116 0.05721002 0.621609825 0.8375784   1
#> var50   14 3 0.7647557 0.05930961 0.649196208 0.8755927   1
#> var51   15 3 0.7561532 0.06015950 0.649150458 0.8874273   1
#> var52   16 3 0.7108223 0.06337808 0.588784788 0.8324290   1
#> var53   17 3 0.7047157 0.05364273 0.598700591 0.8016899   1
#> var54   18 3 0.7099413 0.05282302 0.614491281 0.8091791   1
#> 
#> $qlambda
#>          [,1]      [,2]      [,3]
#> I1  0.7380810 0.0000000 0.0000000
#> I2  0.7770108 0.0000000 0.0000000
#> I3  0.7538982 0.0000000 0.0000000
#> I4  0.6924366 0.0000000 0.0000000
#> I5  0.7519420 0.2289749 0.0000000
#> I6  0.7467335 0.2427417 0.0000000
#> I7  0.0000000 0.7697333 0.0000000
#> I8  0.0000000 0.7689378 0.0000000
#> I9  0.0000000 0.7711148 0.0000000
#> I10 0.0000000 0.6957037 0.0000000
#> I11 0.0000000 0.7367116 0.0000000
#> I12 0.0000000 0.7497902 0.0000000
#> I13 0.0000000 0.0000000 0.7348116
#> I14 0.0000000 0.0000000 0.7647557
#> I15 0.0000000 0.0000000 0.7561532
#> I16 0.0000000 0.0000000 0.7108223
#> I17 0.2731530 0.0000000 0.7047157
#> I18 0.2801235 0.0000000 0.7099413
#> 
#> $eigen
#>         est        sd    lower    upper sig
#> F1 3.585394 0.5650524 2.760307 4.781705   1
#> F2 3.656334 0.7488019 2.462118 5.047279   1
#> F3 3.468136 0.4567995 2.649537 4.390786   1
#> 
#> $dpsx
#>           est         sd     lower     upper sig
#> I1  0.4947443 0.02537462 0.4453227 0.5450595   1
#> I2  0.4747848 0.02407605 0.4234813 0.5165734   1
#> I3  0.4939726 0.02461745 0.4491618 0.5444413   1
#> I4  0.5324665 0.02597862 0.4834027 0.5826305   1
#> I5  0.2054844 0.01290521 0.1825830 0.2333936   1
#> I6  0.2216360 0.01338865 0.1970645 0.2476119   1
#> I7  0.5168719 0.02575559 0.4692324 0.5708618   1
#> I8  0.5144858 0.02619169 0.4688390 0.5690687   1
#> I9  0.4839741 0.02486998 0.4341564 0.5303802   1
#> I10 0.5443913 0.02685778 0.4945853 0.5957864   1
#> I11 0.2306706 0.01380884 0.2027840 0.2553676   1
#> I12 0.2115242 0.01339669 0.1842514 0.2362673   1
#> I13 0.4871659 0.02541251 0.4392566 0.5377029   1
#> I14 0.4947683 0.02522876 0.4460933 0.5459318   1
#> I15 0.4798667 0.02424535 0.4319277 0.5244874   1
#> I16 0.5082189 0.02659450 0.4608719 0.5666127   1
#> I17 0.2258619 0.01316764 0.2025199 0.2539956   1
#> I18 0.2112309 0.01279172 0.1872675 0.2372608   1
#> 
#> $phi
#>      row col       est        sd     lower     upper sig
#> var1   2   1 0.5691858 0.1467611 0.2246361 0.7551931   1
#> var2   3   1 0.5229470 0.1116751 0.3203469 0.7075037   1
#> var3   3   2 0.5707190 0.1152208 0.3676309 0.7669323   1
#> 
#> $shrink
#>          est        sd   lower    upper sig
#> F1  3.381835 0.5067763 2.46928 4.357771   1
#> F2  3.381835 0.5067763 2.46928 4.357771   1
#> F3  3.381835 0.5067763 2.46928 4.357771   1
#> PSX 0.000000 0.0000000 0.00000 0.000000   0
#> 

# }