(Generalized) Partially Confirmatory Factor Analysis

PCFA is a partially confirmatory approach covering a wide range of the exploratory-confirmatory continuum in factor analytic models (Chen, Guo, Zhang, & Pan, 2020). The PCFA is only for continuous data, while the generalized PCFA (GPCFA) covers both continuous and categorical data. There are two major model variants with different constraints for identification. One assumes local independence (LI) with a more exploratory tendency, which can be also called the E-step. The other allows local dependence (LD) with a more confirmatory tendency, which can be also called the C-step. Parameters are obtained by sampling from the posterior distributions with the Markov chain Monte Carlo (MCMC) techniques. Different Bayesian Lasso methods are used to regularize the loading pattern and LD. The estimation results can be summarized with summary.lawbl and the factorial eigenvalue can be plotted with plot_lawbl.

pcfa(
  dat,
  Q,
  LD = TRUE,
  cati = NULL,
  cand_thd = 0.2,
  PPMC = FALSE,
  burn = 5000,
  iter = 5000,
  update = 1000,
  missing = NA,
  rseed = 12345,
  digits = 4,
  alas = FALSE,
  verbose = FALSE
)

Arguments

dat
A \(N \times J\) data matrix or data.frame consisting of the responses of \(N\) individuals to \(J\) items.

Q
A \(J \times K\) design matrix for the loading pattern with \(K\) factors and \(J\) items. Elements are 1, -1, and 0 for specified, unspecified, and zero-fixed loadings, respectively. For models with LI or the E-step, one can specify a few (e.g., 2) loadings per factor. For models with LD or the C-step, the sufficient condition of one specified loading per item is suggested, although there can be a few items without any specified loading. See Examples.

LD
logical; TRUE for allowing LD (model with LD or C-step).

cati
The set of categorical (polytomous) items in sequence number (i.e., 1 to \(J\)); NULL for no and -1 for all items (default is NULL).

cand_thd
Candidate parameter for sampling the thresholds with the MH algorithm.

PPMC
logical; TRUE for conducting posterior predictive model checking.

burn
Number of burn-in iterations before posterior sampling.

iter
Number of formal iterations for posterior sampling (> 0).

update
Number of iterations to update the sampling information.

missing
Value for missing data (default is NA).

rseed
An integer for the random seed.

digits
Number of significant digits to print when printing numeric values.

alas
logical; for adaptive Lasso or not. The default is FALSE.

verbose
logical; to display the sampling information every update or not.

Feigen: Eigenvalue for each factor.

NLA_le3: Number of Loading estimates >= .3 for each factor.

Shrink: Ave. shrinkage parameter (for adaptive LASSO) for each factor.

Adj PSR: Adjusted PSR for each factor.

Ave. Thd: Ave. thresholds.

Acc Rate: Acceptance rate of threshold (MH algorithm).

LD>.2 >.1 LD>.2 >.1: # of LD terms larger than .2 and .1, and LD's shrinkage parameter.

Value

pcfa returns an object of class lawbl without item intercepts. It contains a lot of information about the posteriors that can be summarized using summary.lawbl.

References

Chen, J., Guo, Z., Zhang, L., & Pan, J. (2020). A partially confirmatory approach to scale development with the Bayesian Lasso. Psychological Methods. Advance online publication. DOI: 10.1037/met0000293.

Examples

# \donttest{
#####################################################
#  Example 1: Estimation with continuous data & LD  #
#####################################################

dat <- sim18cfa1$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:6,1]<-Q[7:12,2]<-Q[13:18,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = TRUE,burn = 2000, iter = 2000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] TRUE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`True Factor`
#> [1] TRUE TRUE TRUE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.267631   1.862111
#> F2   1.133440   1.479512
#> F3   1.717687   3.043729
#> 
#> $`No. of sig LD terms`
#> [1] 6
#> 

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.6777371 0.0000000 0.0000000
#> I2  0.6947720 0.0000000 0.0000000
#> I3  0.7287455 0.0000000 0.0000000
#> I4  0.7283320 0.0000000 0.0000000
#> I5  0.7083776 0.3824132 0.0000000
#> I6  0.6959418 0.3621274 0.0000000
#> I7  0.0000000 0.6990281 0.0000000
#> I8  0.0000000 0.7370763 0.0000000
#> I9  0.0000000 0.7200735 0.0000000
#> I10 0.0000000 0.7271835 0.0000000
#> I11 0.0000000 0.6997877 0.3486253
#> I12 0.0000000 0.7010905 0.3532886
#> I13 0.0000000 0.0000000 0.7224160
#> I14 0.0000000 0.0000000 0.6943947
#> I15 0.0000000 0.0000000 0.7263602
#> I16 0.0000000 0.0000000 0.7069214
#> I17 0.3596368 0.0000000 0.7175995
#> I18 0.3393079 0.0000000 0.7264559

summary(m0, what = 'offpsx') #summarize significant LD terms

#>        row col       est         sd     lower     upper sig
#> var14   14   1 0.2647726 0.03642568 0.1926225 0.3415343   1
#> var24    7   2 0.3108189 0.03578501 0.2415014 0.3850120   1
#> var37    4   3 0.2946160 0.04782329 0.2053247 0.3892355   1
#> var111  13   8 0.2554767 0.05178663 0.1549109 0.3584666   1
#> var118  10   9 0.2750974 0.04487471 0.1931764 0.3664781   1
#> var163  16  15 0.3032920 0.05100715 0.2110932 0.4014631   1

######################################################
#  Example 2: Estimation with categorical data & LI  #
######################################################
dat <- sim18ccfa40$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[7:8,2]<-Q[13:14,3]<-1

m1 <- pcfa(dat = dat, Q = Q,LD = FALSE,cati=-1,burn = 2000, iter = 2000)
summary(m1) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 9.883333
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`True Factor`
#> [1] TRUE TRUE TRUE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.036511   1.046103
#> F2   1.190794   1.485645
#> F3   2.520378   6.890709
#> 
#> $`Cat Items`
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18
#> 
#> $`max No. of categories`
#> [1] 4
#> 

summary(m1, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7598362 0.0000000 0.0000000
#> I2  0.7631264 0.0000000 0.0000000
#> I3  0.7110248 0.0000000 0.0000000
#> I4  0.6672460 0.0000000 0.0000000
#> I5  0.7296244 0.4397355 0.0000000
#> I6  0.7009867 0.4111027 0.0000000
#> I7  0.0000000 0.7503882 0.0000000
#> I8  0.0000000 0.6886856 0.0000000
#> I9  0.0000000 0.6915755 0.0000000
#> I10 0.0000000 0.7430489 0.0000000
#> I11 0.0000000 0.7189176 0.3350080
#> I12 0.0000000 0.7036005 0.3837379
#> I13 0.0000000 0.0000000 0.7372507
#> I14 0.0000000 0.0000000 0.7213758
#> I15 0.0000000 0.0000000 0.7313625
#> I16 0.0000000 0.0000000 0.7163863
#> I17 0.2583261 0.0000000 0.7718187
#> I18 0.2549616 0.0000000 0.7676993

summary(m1, what = 'offpsx') #summarize significant LD terms

#> NULL

summary(m1,what='thd') #thresholds for categorical items

#>          [,1]         [,2]     [,3]
#> I1  -1.521470 -0.033597986 1.511013
#> I2  -1.549058 -0.019421984 1.616661
#> I3  -1.584611 -0.071919752 1.455610
#> I4  -1.467652 -0.088016543 1.460645
#> I5  -1.660025 -0.068914279 1.476852
#> I6  -1.574076 -0.026917267 1.509269
#> I7  -1.493155  0.050588533 1.690817
#> I8  -1.523947  0.018727926 1.404398
#> I9  -1.499463  0.017813981 1.539265
#> I10 -1.455960 -0.020035736 1.480965
#> I11 -1.647587 -0.022561448 1.543301
#> I12 -1.593390  0.005463383 1.512301
#> I13 -1.392118  0.053354389 1.566341
#> I14 -1.439920  0.057642536 1.601957
#> I15 -1.401346  0.041512689 1.446094
#> I16 -1.476899  0.040457459 1.560623
#> I17 -1.493393 -0.061548455 1.562806
#> I18 -1.535266 -0.056733222 1.481305

# }