Partially Confirmatory Factor Analysis

PCFA is a partially confirmatory approach covering a wide range of the exploratory-confirmatory continuum in factor analytic models (Chen, Guo, Zhang, & Pan, 2020). There are two major model variants with different constraints for identification. One assumes local independence (LI) with a more exploratory tendency, which can be also called the E-step. The other allows local dependence (LD) with a more confirmatory tendency, which can be also called the C-step. Parameters are obtained by sampling from the posterior distributions with the Markov chain Monte Carlo (MCMC) techniques. Different Bayesian Lasso methods are used to regularize the loading pattern and LD. The estimation results can be summarized with summary.lawbl and the factorial eigenvalue can be plotted with plot_eigen.

pcfa(
  dat,
  Q,
  LD = TRUE,
  PPMC = FALSE,
  burn = 5000,
  iter = 5000,
  update = 1000,
  missing = NA,
  rseed = 12345,
  digits = 4,
  alas = FALSE
)

Arguments

dat
A \(N \times J\) matrix or data.frame consisting of the responses of \(N\) individuals to \(J\) items. Only continuous data are supported currently.

Q
A \(J \times K\) design matrix for the loading pattern with \(K\) factors and \(J\) items. Elements are 1, -1, and 0 for specified, unspecified, and zero-fixed loadings, respectively. For models with LI or the E-step, one can specify a few (e.g., 2) loadings per factor. For models with LD or the C-step, the sufficient condition of one specified loading per item is suggested, although there can be a few items without any specified loading (Chen, Guo, Zhang, & Pan, 2020). See Examples.

LD
logical; TRUE for allowing LD (model with LD or C-step).

PPMC
logical; TRUE for conducting posterior predictive model checking.

burn
Number of burn-in iterations before posterior sampling.

iter
Number of formal iterations for posterior sampling.

update
Number of iterations to update the sampling information.

missing
Value for missing data (default is NA) (under development).

rseed
An integer for the random seed.

digits
Number of significant digits to print when printing numeric values.

alas
logical; for adaptive Lasso or not. The default is FALSE, which seems slightly stabler.

Value

pcfa returns an object of class pcfa. It contains a lot of information about the posteriors that can be summarized using summary.lawbl. The factorial eigenvalue can be plotted with plot_eigen.

References

Chen, J., Guo, Z., Zhang, L., & Pan, J. (2020). A partially confirmatory approach to scale development with the Bayesian Lasso. Psychological Methods. Advance online publication. DOI:10.1037/met0000293

Examples

# \donttest{
####################################
#  Example 1: Estimation with LI   #
####################################

dat <- sim18cfa1$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1
Q

#>       [,1] [,2] [,3]
#>  [1,]    1   -1   -1
#>  [2,]    1   -1   -1
#>  [3,]   -1   -1   -1
#>  [4,]   -1   -1   -1
#>  [5,]   -1   -1   -1
#>  [6,]   -1   -1   -1
#>  [7,]   -1   -1   -1
#>  [8,]   -1   -1   -1
#>  [9,]   -1    1   -1
#> [10,]   -1    1   -1
#> [11,]   -1   -1   -1
#> [12,]   -1   -1   -1
#> [13,]   -1   -1    1
#> [14,]   -1   -1    1
#> [15,]   -1   -1   -1
#> [16,]   -1   -1   -1
#> [17,]   -1   -1   -1
#> [18,]   -1   -1   -1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 500
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 23
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.240815   1.809868
#> F2   1.287728   1.986893
#> F3   1.211756   1.763918
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd       lower     upper sig
#> var1     1 1 0.6322639 0.07648500 0.481045840 0.7781675   1
#> var2     2 1 0.6600763 0.08347573 0.497903739 0.8206087   1
#> var3     3 1 0.9025356 0.08491108 0.732804521 1.0638966   1
#> var4     4 1 0.9092650 0.08542165 0.760099113 1.0773152   1
#> var5     5 1 0.6355668 0.12142862 0.383291632 0.8436551   1
#> var6     6 1 0.6566113 0.12349769 0.397996588 0.8800632   1
#> var18   18 1 0.2820228 0.15010748 0.002326382 0.5428930   1
#> var23    5 2 0.3947471 0.09546164 0.223472825 0.6126997   1
#> var24    6 2 0.4025233 0.09829888 0.179455453 0.5832987   1
#> var25    7 2 0.7304012 0.06436148 0.602318986 0.8503856   1
#> var26    8 2 0.6341750 0.05849776 0.524353406 0.7521323   1
#> var27    9 2 0.8196437 0.06108358 0.712873523 0.9432036   1
#> var28   10 2 0.8225157 0.06441270 0.708728612 0.9554128   1
#> var29   11 2 0.6958852 0.06260122 0.562023190 0.8200892   1
#> var30   12 2 0.6997449 0.06041724 0.578132343 0.8303693   1
#> var47   11 3 0.3852286 0.09464164 0.223164972 0.5873377   1
#> var48   12 3 0.3752705 0.09405357 0.203586768 0.5749404   1
#> var49   13 3 0.7241304 0.09033567 0.564313467 0.9012510   1
#> var50   14 3 0.7225850 0.08920329 0.580112357 0.9107763   1
#> var51   15 3 0.9307828 0.09867094 0.760426682 1.1160452   1
#> var52   16 3 0.8874406 0.09539132 0.735708650 1.0781340   1
#> var53   17 3 0.7314842 0.10422864 0.563920302 0.9395408   1
#> var54   18 3 0.7256233 0.10771381 0.544781401 0.9377102   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.6322639 0.0000000 0.0000000
#> I2  0.6600763 0.0000000 0.0000000
#> I3  0.9025356 0.0000000 0.0000000
#> I4  0.9092650 0.0000000 0.0000000
#> I5  0.6355668 0.3947471 0.0000000
#> I6  0.6566113 0.4025233 0.0000000
#> I7  0.0000000 0.7304012 0.0000000
#> I8  0.0000000 0.6341750 0.0000000
#> I9  0.0000000 0.8196437 0.0000000
#> I10 0.0000000 0.8225157 0.0000000
#> I11 0.0000000 0.6958852 0.3852286
#> I12 0.0000000 0.6997449 0.3752705
#> I13 0.0000000 0.0000000 0.7241304
#> I14 0.0000000 0.0000000 0.7225850
#> I15 0.0000000 0.0000000 0.9307828
#> I16 0.0000000 0.0000000 0.8874406
#> I17 0.0000000 0.0000000 0.7314842
#> I18 0.2820228 0.0000000 0.7256233

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

####################################
#  Example 2: Estimation with LD   #
####################################

Q<-matrix(-1,J,K);
Q[1:6,1]<-Q[7:12,2]<-Q[13:18,3]<-1
Q

#>       [,1] [,2] [,3]
#>  [1,]    1   -1   -1
#>  [2,]    1   -1   -1
#>  [3,]    1   -1   -1
#>  [4,]    1   -1   -1
#>  [5,]    1   -1   -1
#>  [6,]    1   -1   -1
#>  [7,]   -1    1   -1
#>  [8,]   -1    1   -1
#>  [9,]   -1    1   -1
#> [10,]   -1    1   -1
#> [11,]   -1    1   -1
#> [12,]   -1    1   -1
#> [13,]   -1   -1    1
#> [14,]   -1   -1    1
#> [15,]   -1   -1    1
#> [16,]   -1   -1    1
#> [17,]   -1   -1    1
#> [18,]   -1   -1    1

m1 <- pcfa(dat = dat, Q = Q,burn = 1000, iter = 1000)
summary(m1) # summarize basic information

#> $N
#> [1] 500
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] TRUE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.070743   1.230802
#> F2   1.110975   1.394687
#> F3   1.009564   1.034085
#> 
#> $`No. of sig LD terms`
#> [1] 6
#> 

summary(m1, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7509411 0.0000000 0.0000000
#> I2  0.6813936 0.0000000 0.0000000
#> I3  0.7922422 0.0000000 0.0000000
#> I4  0.7689045 0.0000000 0.0000000
#> I5  0.6728831 0.3473103 0.0000000
#> I6  0.6919241 0.3526614 0.0000000
#> I7  0.0000000 0.7569042 0.0000000
#> I8  0.0000000 0.6349059 0.0000000
#> I9  0.0000000 0.7406121 0.0000000
#> I10 0.0000000 0.7418163 0.0000000
#> I11 0.0000000 0.7341822 0.3399042
#> I12 0.0000000 0.7426154 0.3251054
#> I13 0.0000000 0.0000000 0.7107169
#> I14 0.0000000 0.0000000 0.6884674
#> I15 0.0000000 0.0000000 0.8007282
#> I16 0.0000000 0.0000000 0.7607961
#> I17 0.3172098 0.0000000 0.7127242
#> I18 0.3257416 0.0000000 0.7231049

summary(m1, what = 'offpsx') #summarize significant LD terms

#>        row col       est         sd     lower     upper sig
#> var14   14   1 0.2710355 0.04118050 0.1910537 0.3532721   1
#> var24    7   2 0.2531845 0.04669585 0.1633123 0.3464047   1
#> var37    4   3 0.2839852 0.06403869 0.1697131 0.4098167   1
#> var111  13   8 0.3046336 0.04292341 0.2246645 0.3912058   1
#> var118  10   9 0.2594980 0.05428821 0.1520387 0.3737294   1
#> var163  16  15 0.2309577 0.05004079 0.1356811 0.3282657   1

# }