Partially Confirmatory Item Response Model

pcirm is a partially confirmatory approach to item response models (Chen, 2020), which estimates the intercept for continuous and dichotomous data. Similar to PCFA and GPCFA, there are two major model variants with different constraints for identification. One assumes local independence (LI) with a more exploratory tendency, which can be also called the E-step. The other allows local dependence (LD) with a more confirmatory tendency, which can be also called the C-step. Parameters are obtained by sampling from the posterior distributions with the Markov chain Monte Carlo (MCMC) techniques. Different Bayesian Lasso methods are used to regularize the loading pattern and LD. The estimation results can be summarized with summary.lawbl and the factorial eigenvalue can be plotted with plot_eigen.

pcirm(
  dat,
  Q,
  LD = TRUE,
  cati = NULL,
  PPMC = FALSE,
  burn = 5000,
  iter = 5000,
  update = 1000,
  missing = NA,
  rseed = 12345,
  digits = 4,
  alas = FALSE,
  verbose = FALSE
)

Arguments

dat
A \(N \times J\) data matrix or data.frame consisting of the responses of \(N\) individuals to \(J\) items. Only continuous and dichotomous data are supported.

Q
A \(J \times K\) design matrix for the loading pattern with \(K\) factors and \(J\) items. Elements are 1, -1, and 0 for specified, unspecified, and zero-fixed loadings, respectively. For models with LI or the E-step, one can specify a few (e.g., 2) loadings per factor. For models with LD or the C-step, the sufficient condition of one specified loading per item is suggested, although there can be a few items without any specified loading. See Examples.

LD
logical; TRUE for allowing LD (model with LD or C-step).

cati
The set of dichotomous items in sequence number (i.e., 1 to \(J\)); NULL for no and -1 for all items (default is NULL).

PPMC
logical; TRUE for conducting posterior predictive model checking.

burn
Number of burn-in iterations before posterior sampling.

iter
Number of formal iterations for posterior sampling (> 0).

update
Number of iterations to update the sampling information.

missing
Value for missing data (default is NA).

rseed
An integer for the random seed.

digits
Number of significant digits to print when printing numeric values.

alas
logical; for adaptive Lasso or not. The default is FALSE.

verbose
logical; to display the sampling information every update or not.

Feigen: Eigenvalue for each factor.

NLA_lg3: Number of Loading estimates > .3 for each factor.

Shrink: Ave. shrinkage parameter (for adaptive LASSO) for each factor.

Adj PSR: Adjusted PSR for each factor.

Ave. Int.: Ave. item intercept.

LD>.2 >.1 LD>.2 >.1: # of LD terms larger than .2 and .1, and LD's shrinkage parameter.

Value

pcirm returns an object of class lawbl with item intercepts. It contains a lot of information about the posteriors that can be summarized using summary.lawbl. The factorial eigenvalue can be plotted with plot_eigen.

References

Chen, J. (2020). A partially confirmatory approach to the multidimensional item response theory with the Bayesian Lasso. Psychometrika. 85(3), 738-774. DOI:10.1007/s11336-020-09724-3.

Examples

# \donttest{
####################################
#  Example 1: Estimation with LD   #
####################################

dat <- sim24ccfa21$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:8,1]<-Q[9:16,2]<-Q[17:24,3]<-1

m0 <- pcirm(dat = dat, Q = Q, LD = TRUE, cati = -1, burn = 2000,iter = 2000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 24
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] TRUE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.007454   1.037260
#> F2   1.018612   1.041351
#> F3   1.048443   1.183139
#> 
#> $`No. of sig LD terms`
#> [1] 7
#> 
#> $`Cat Items`
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> 
#> $`max No. of categories`
#> [1] 2
#> 

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.6445135 0.0000000 0.0000000
#> I2  0.6441112 0.0000000 0.0000000
#> I3  0.6295697 0.0000000 0.0000000
#> I4  0.6781843 0.0000000 0.0000000
#> I5  0.6486247 0.0000000 0.0000000
#> I6  0.6570034 0.0000000 0.0000000
#> I7  0.7456747 0.0000000 0.0000000
#> I8  0.7766815 0.0000000 0.0000000
#> I9  0.0000000 0.7257810 0.0000000
#> I10 0.0000000 0.6927725 0.0000000
#> I11 0.0000000 0.7565709 0.0000000
#> I12 0.0000000 0.6967574 0.0000000
#> I13 0.0000000 0.7290665 0.0000000
#> I14 0.0000000 0.6948335 0.0000000
#> I15 0.0000000 0.6341475 0.0000000
#> I16 0.0000000 0.6751562 0.0000000
#> I17 0.0000000 0.0000000 0.6924022
#> I18 0.0000000 0.0000000 0.6824835
#> I19 0.0000000 0.0000000 0.7199200
#> I20 0.0000000 0.0000000 0.6880398
#> I21 0.0000000 0.0000000 0.7582321
#> I22 0.0000000 0.0000000 0.6672866
#> I23 0.0000000 0.0000000 0.6868795
#> I24 0.0000000 0.0000000 0.6839064

summary(m0, what = 'offpsx') #summarize significant LD terms

#>        row col       est         sd      lower     upper sig
#> var131   8   7 0.1225632 0.05107477 0.02149781 0.2150973   1
#> var247  16  15 0.1619993 0.04495219 0.07745951 0.2520529   1
#> var254  23  15 0.1508254 0.03976499 0.07417408 0.2296566   1
#> var255  24  15 0.1402446 0.04257192 0.06312714 0.2318276   1
#> var263  23  16 0.1524144 0.04024589 0.06990800 0.2307790   1
#> var264  24  16 0.1448762 0.04121645 0.06604447 0.2266492   1
#> var299  24  23 0.1634590 0.04990897 0.08109708 0.2666170   1

####################################
#  Example 2: Estimation with LD   #
####################################

Q<-cbind(Q,-1);
Q[15:16,4]<-1

m1 <- pcirm(dat = dat, Q = Q, LD = FALSE, cati = -1, burn = 2000,iter = 2000)
summary(m1) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 24
#> 
#> $K
#> [1] 4
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 2000
#> 
#> $Iteration
#> [1] 2000
#> 
#> $`No. of sig lambda`
#> [1] 28
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.045718   1.064053
#> F2   1.193415   1.657691
#> F3   1.065911   1.174982
#> F4   1.123002   1.423497
#> 
#> $`Cat Items`
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
#> 
#> $`max No. of categories`
#> [1] 2
#> 

summary(m1, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]      [,4]
#> I1  0.5962339 0.0000000 0.0000000 0.0000000
#> I2  0.6105347 0.0000000 0.0000000 0.0000000
#> I3  0.6271678 0.0000000 0.0000000 0.0000000
#> I4  0.6958273 0.0000000 0.0000000 0.0000000
#> I5  0.6114661 0.0000000 0.0000000 0.0000000
#> I6  0.6154533 0.0000000 0.0000000 0.0000000
#> I7  0.8689566 0.0000000 0.0000000 0.0000000
#> I8  0.8939026 0.0000000 0.0000000 0.0000000
#> I9  0.0000000 0.7577736 0.0000000 0.0000000
#> I10 0.0000000 0.6824360 0.0000000 0.0000000
#> I11 0.0000000 0.7546485 0.0000000 0.0000000
#> I12 0.0000000 0.6966081 0.0000000 0.0000000
#> I13 0.0000000 0.7472792 0.0000000 0.0000000
#> I14 0.0000000 0.7009093 0.0000000 0.0000000
#> I15 0.0000000 0.5692394 0.0000000 0.5656617
#> I16 0.0000000 0.6135148 0.0000000 0.5681055
#> I17 0.0000000 0.0000000 0.7140304 0.0000000
#> I18 0.0000000 0.0000000 0.6982343 0.0000000
#> I19 0.0000000 0.0000000 0.7111138 0.0000000
#> I20 0.0000000 0.0000000 0.7153510 0.0000000
#> I21 0.0000000 0.0000000 0.7582782 0.0000000
#> I22 0.0000000 0.0000000 0.6833458 0.0000000
#> I23 0.0000000 0.0000000 0.6288027 0.4751283
#> I24 0.0000000 0.0000000 0.6422785 0.4371894

summary(m1, what = 'offpsx') #summarize significant LD terms

#> NULL

# }