Simulating data with Latent Variable Modeling

sim_lvm can simulate data based on factor analysis or item response models with different response formats (continuous or categorical), loading patterns and residual covariance (local dependence) structures.

sim_lvm(
  N = 1000,
  K = 3,
  ipf = 8,
  cpf = 2,
  lam = 0.7,
  lac = 0.3,
  phi = 0.5,
  ph1 = -1,
  ecr = 0,
  ome_out = FALSE,
  cati = NULL,
  noc = c(4),
  misp = 0,
  rseed = 333,
  necw = K,
  necb = K,
  add_ind = c(),
  add_la = 0.5,
  add_phi = 0,
  zero_it = 0,
  digits = 4
)

Arguments

N
Sample size.

K
Number of factors.

ipf
Items per factor.

cpf
Cross-loadings per factor.

lam
Number of formal iterations for posterior sampling.

lac
Number of iterations to update the sampling information.

phi
Homogeneous correlations between any two factors.

ph1
Correlation between factor 1 and 2 (if it's different from phi.

ecr
Residual correlation (local dependence).

ome_out
Output factor score or not.

cati
The set of categorical (polytomous) items in sequence number (i.e., 1 to \(J\)); NULL for no and -1 for all (default is NULL).

noc
Number of categories for categorical items

misp
Proportion of missingness.

rseed
An integer for the random seed.

necw
Number of within-factor local dependence.

necb
Number of between-factor local dependence.

add_ind
(Additional) minor factor with cross-loadings.

add_la
Value of cross-loadings on (Additional) minor factor.

add_phi
Correlations between (Additional) minor factor and other factors.

zero_it
Surplus items with zero loading.

digits
Number of significant digits to print when printing numeric values.

Value

An object of class list containing the data, loading, and factorial correlation matrix.

Examples

# for continuous data with cross-loadings and local dependence effect .3
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.3,ecr=.3)
summary(out$dat)

#>        V1                 V2                 V3                  V4          
#>  Min.   :-3.46922   Min.   :-3.28985   Min.   :-3.378121   Min.   :-3.21486  
#>  1st Qu.:-0.65125   1st Qu.:-0.68834   1st Qu.:-0.626712   1st Qu.:-0.68908  
#>  Median :-0.02374   Median :-0.04829   Median :-0.037197   Median :-0.02832  
#>  Mean   :-0.04079   Mean   :-0.02890   Mean   : 0.003397   Mean   :-0.01200  
#>  3rd Qu.: 0.55393   3rd Qu.: 0.62665   3rd Qu.: 0.660568   3rd Qu.: 0.67534  
#>  Max.   : 3.14863   Max.   : 3.17885   Max.   : 3.971842   Max.   : 2.85749  
#>        V5                  V6                 V7                 V8          
#>  Min.   :-3.927899   Min.   :-3.90691   Min.   :-3.09723   Min.   :-3.52391  
#>  1st Qu.:-0.667047   1st Qu.:-0.67925   1st Qu.:-0.65529   1st Qu.:-0.53677  
#>  Median : 0.008415   Median : 0.01308   Median : 0.02744   Median : 0.06098  
#>  Mean   : 0.013330   Mean   : 0.01076   Mean   : 0.03409   Mean   : 0.07785  
#>  3rd Qu.: 0.677886   3rd Qu.: 0.70534   3rd Qu.: 0.68127   3rd Qu.: 0.78287  
#>  Max.   : 2.792838   Max.   : 3.16835   Max.   : 3.71842   Max.   : 3.11273  
#>        V9                V10                V11                 V12          
#>  Min.   :-3.75124   Min.   :-3.09726   Min.   :-3.681806   Min.   :-2.94354  
#>  1st Qu.:-0.68644   1st Qu.:-0.70762   1st Qu.:-0.658355   1st Qu.:-0.60576  
#>  Median :-0.02573   Median :-0.04845   Median :-0.002415   Median : 0.06436  
#>  Mean   :-0.02954   Mean   :-0.02514   Mean   : 0.027651   Mean   : 0.03694  
#>  3rd Qu.: 0.66967   3rd Qu.: 0.65978   3rd Qu.: 0.707049   3rd Qu.: 0.69576  
#>  Max.   : 2.88798   Max.   : 3.74679   Max.   : 2.944017   Max.   : 2.82441  
#>       V13                V14                V15                V16          
#>  Min.   :-3.09344   Min.   :-3.63601   Min.   :-3.34665   Min.   :-3.78474  
#>  1st Qu.:-0.63972   1st Qu.:-0.70639   1st Qu.:-0.68965   1st Qu.:-0.60530  
#>  Median : 0.08316   Median :-0.01224   Median : 0.04764   Median : 0.06889  
#>  Mean   : 0.04668   Mean   : 0.01340   Mean   : 0.01542   Mean   : 0.03929  
#>  3rd Qu.: 0.72967   3rd Qu.: 0.72401   3rd Qu.: 0.75258   3rd Qu.: 0.74196  
#>  Max.   : 3.06067   Max.   : 2.95634   Max.   : 3.03290   Max.   : 3.41078  
#>       V17                 V18           
#>  Min.   :-2.963458   Min.   :-3.488045  
#>  1st Qu.:-0.669276   1st Qu.:-0.670281  
#>  Median :-0.024789   Median :-0.011018  
#>  Mean   : 0.002917   Mean   : 0.004546  
#>  3rd Qu.: 0.675210   3rd Qu.: 0.663606  
#>  Max.   : 3.303036   Max.   : 3.011866  

out$MLA

#>       [,1] [,2] [,3]
#>  [1,]  0.7  0.0  0.0
#>  [2,]  0.7  0.0  0.0
#>  [3,]  0.7  0.0  0.0
#>  [4,]  0.7  0.0  0.0
#>  [5,]  0.7  0.3  0.0
#>  [6,]  0.7  0.3  0.0
#>  [7,]  0.0  0.7  0.0
#>  [8,]  0.0  0.7  0.0
#>  [9,]  0.0  0.7  0.0
#> [10,]  0.0  0.7  0.0
#> [11,]  0.0  0.7  0.3
#> [12,]  0.0  0.7  0.3
#> [13,]  0.0  0.0  0.7
#> [14,]  0.0  0.0  0.7
#> [15,]  0.0  0.0  0.7
#> [16,]  0.0  0.0  0.7
#> [17,]  0.3  0.0  0.7
#> [18,]  0.3  0.0  0.7

out$ofd_ind

#>      row col
#> [1,]  14   1
#> [2,]   7   2
#> [3,]   4   3
#> [4,]  13   8
#> [5,]  10   9
#> [6,]  16  15

# for categorical data with cross-loadings .4 and 10% missingness
out <- sim_lvm(N=1000,K=3,ipf=6,lam = .7, lac=.4,cati=-1,noc=4,misp=.1)
summary(out$dat)

#>        V1              V2              V3              V4       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :2.000   Median :2.000   Median :2.000   Median :2.000  
#>  Mean   :2.507   Mean   :2.508   Mean   :2.498   Mean   :2.464  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000  
#>  Max.   :4.000   Max.   :4.000   Max.   :4.000   Max.   :4.000  
#>  NA's   :91      NA's   :92      NA's   :93      NA's   :104    
#>        V5              V6              V7              V8       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :3.000   Median :2.000   Median :3.000   Median :3.000  
#>  Mean   :2.513   Mean   :2.499   Mean   :2.511   Mean   :2.537  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000  
#>  Max.   :4.000   Max.   :4.000   Max.   :4.000   Max.   :4.000  
#>  NA's   :105     NA's   :119     NA's   :115     NA's   :108    
#>        V9             V10             V11             V12            V13       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.00   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.00   1st Qu.:2.000  
#>  Median :3.000   Median :3.000   Median :3.000   Median :3.00   Median :2.000  
#>  Mean   :2.507   Mean   :2.518   Mean   :2.528   Mean   :2.52   Mean   :2.459  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.00   3rd Qu.:3.000  
#>  Max.   :4.000   Max.   :4.000   Max.   :4.000   Max.   :4.00   Max.   :4.000  
#>  NA's   :102     NA's   :99      NA's   :92      NA's   :100    NA's   :103    
#>       V14             V15             V16             V17       
#>  Min.   :1.000   Min.   :1.000   Min.   :1.000   Min.   :1.000  
#>  1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000   1st Qu.:2.000  
#>  Median :3.000   Median :3.000   Median :3.000   Median :2.000  
#>  Mean   :2.544   Mean   :2.501   Mean   :2.508   Mean   :2.496  
#>  3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000   3rd Qu.:3.000  
#>  Max.   :4.000   Max.   :4.000   Max.   :4.000   Max.   :4.000  
#>  NA's   :96      NA's   :114     NA's   :96      NA's   :84     
#>       V18       
#>  Min.   :1.000  
#>  1st Qu.:2.000  
#>  Median :2.000  
#>  Mean   :2.479  
#>  3rd Qu.:3.000  
#>  Max.   :4.000  
#>  NA's   :110    

out$MLA

#>       [,1] [,2] [,3]
#>  [1,]  0.7  0.0  0.0
#>  [2,]  0.7  0.0  0.0
#>  [3,]  0.7  0.0  0.0
#>  [4,]  0.7  0.0  0.0
#>  [5,]  0.7  0.4  0.0
#>  [6,]  0.7  0.4  0.0
#>  [7,]  0.0  0.7  0.0
#>  [8,]  0.0  0.7  0.0
#>  [9,]  0.0  0.7  0.0
#> [10,]  0.0  0.7  0.0
#> [11,]  0.0  0.7  0.4
#> [12,]  0.0  0.7  0.4
#> [13,]  0.0  0.0  0.7
#> [14,]  0.0  0.0  0.7
#> [15,]  0.0  0.0  0.7
#> [16,]  0.0  0.0  0.7
#> [17,]  0.4  0.0  0.7
#> [18,]  0.4  0.0  0.7

out$ofd_ind

#>      row col