Summary method for lawbl objects

Provide summaries of posterior information for a lawbl object, .

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

thd: Threshold estimates.

int: Intercept estimates (for pcirm only).

shrink: (Ave) shrinkage for each factor's loadings and LD (if LD in pcfa = T).

factor: Are the factors true of not (for EFA).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

# \donttest{
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[7:8,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 24
#> 
#> $`True Factor`
#> [1] TRUE TRUE TRUE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.946765   3.653873
#> F2   1.061199   1.240003
#> F3   1.528924   2.605051
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F        est         sd       lower       upper sig
#> var1     1 1  0.6960698 0.05685614  0.58530276  0.80380356   1
#> var2     2 1  0.7114862 0.05457511  0.61014588  0.82065312   1
#> var3     3 1  0.7164676 0.05798085  0.60436351  0.83106826   1
#> var4     4 1  0.7398584 0.05619135  0.63706501  0.85348284   1
#> var5     5 1  0.7213999 0.06079190  0.60525895  0.83694075   1
#> var6     6 1  0.6939082 0.06066803  0.57171904  0.80061070   1
#> var16   16 1 -0.1623904 0.09823873 -0.36010717 -0.01202933   1
#> var17   17 1  0.2069553 0.10182882  0.02558243  0.36739783   1
#> var23    5 2  0.2622927 0.10594836  0.06448453  0.41884585   1
#> var24    6 2  0.2798563 0.10167906  0.09258888  0.43707822   1
#> var25    7 2  0.7242954 0.04171414  0.64894846  0.81227265   1
#> var26    8 2  0.7217194 0.04245308  0.64695678  0.81433015   1
#> var27    9 2  0.7328375 0.04121716  0.65312303  0.80751729   1
#> var28   10 2  0.7128629 0.04086035  0.63597812  0.79407955   1
#> var29   11 2  0.7216622 0.05217712  0.60624956  0.82881286   1
#> var30   12 2  0.7109634 0.04880895  0.59457772  0.79573816   1
#> var47   11 3  0.2713489 0.07711220  0.12553449  0.41576553   1
#> var48   12 3  0.2558693 0.07606551  0.11716674  0.40095630   1
#> var49   13 3  0.7176707 0.05241258  0.61024395  0.81238978   1
#> var50   14 3  0.6944668 0.05456705  0.59475718  0.80233424   1
#> var51   15 3  0.7287674 0.05633182  0.62655147  0.83706041   1
#> var52   16 3  0.7261280 0.05805350  0.62289840  0.84074159   1
#> var53   17 3  0.7199577 0.05551577  0.62099014  0.82545456   1
#> var54   18 3  0.7248401 0.05621132  0.62177675  0.82678966   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>           [,1]      [,2]      [,3]
#> I1   0.6960698 0.0000000 0.0000000
#> I2   0.7114862 0.0000000 0.0000000
#> I3   0.7164676 0.0000000 0.0000000
#> I4   0.7398584 0.0000000 0.0000000
#> I5   0.7213999 0.2622927 0.0000000
#> I6   0.6939082 0.2798563 0.0000000
#> I7   0.0000000 0.7242954 0.0000000
#> I8   0.0000000 0.7217194 0.0000000
#> I9   0.0000000 0.7328375 0.0000000
#> I10  0.0000000 0.7128629 0.0000000
#> I11  0.0000000 0.7216622 0.2713489
#> I12  0.0000000 0.7109634 0.2558693
#> I13  0.0000000 0.0000000 0.7176707
#> I14  0.0000000 0.0000000 0.6944668
#> I15  0.0000000 0.0000000 0.7287674
#> I16 -0.1623904 0.0000000 0.7261280
#> I17  0.2069553 0.0000000 0.7199577
#> I18  0.0000000 0.0000000 0.7248401

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

# }