Summary method for pcfa objectects

Provide basic information for an PCFA object, and summarize various posteriors.

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

thd: Threshold estimates.

int: Intercept estimates (for pcirm only).

shrink: (Ave) shrinkage for each factor's loadings and LD (if LD in pcfa = T).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

# \donttest{
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[7:8,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 20
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.786925   3.860107
#> F2   1.216101   1.725925
#> F3   1.531601   3.100957
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd      lower     upper sig
#> var1     1 1 0.7389272 0.08509317 0.58605278 0.9108594   1
#> var2     2 1 0.7798104 0.08671655 0.63621637 0.9566822   1
#> var3     3 1 0.7546501 0.08632634 0.61599997 0.9539889   1
#> var4     4 1 0.6964697 0.07836373 0.56402424 0.8644828   1
#> var5     5 1 0.7537004 0.08639961 0.59832343 0.9096751   1
#> var6     6 1 0.7528286 0.08371548 0.60373188 0.9032052   1
#> var25    7 2 0.7739907 0.07989901 0.62839370 0.9282863   1
#> var26    8 2 0.7799826 0.07900484 0.64040462 0.9313014   1
#> var27    9 2 0.7660520 0.07525168 0.64074251 0.9225442   1
#> var28   10 2 0.7026166 0.07250606 0.57897172 0.8485507   1
#> var29   11 2 0.7826122 0.06751288 0.65742989 0.9149613   1
#> var30   12 2 0.7957693 0.07105310 0.66369582 0.9262337   1
#> var47   11 3 0.1824270 0.10984032 0.02338258 0.4153187   1
#> var48   12 3 0.1877502 0.11083291 0.03312712 0.4319404   1
#> var49   13 3 0.7059281 0.09247914 0.55843858 0.8936570   1
#> var50   14 3 0.7378874 0.09698283 0.59122988 0.9329396   1
#> var51   15 3 0.7227175 0.09685810 0.58018214 0.9251605   1
#> var52   16 3 0.6789860 0.09316008 0.54372650 0.8798640   1
#> var53   17 3 0.6725952 0.09647909 0.52415897 0.8704844   1
#> var54   18 3 0.6753587 0.09590525 0.54306683 0.8866462   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7389272 0.0000000 0.0000000
#> I2  0.7798104 0.0000000 0.0000000
#> I3  0.7546501 0.0000000 0.0000000
#> I4  0.6964697 0.0000000 0.0000000
#> I5  0.7537004 0.0000000 0.0000000
#> I6  0.7528286 0.0000000 0.0000000
#> I7  0.0000000 0.7739907 0.0000000
#> I8  0.0000000 0.7799826 0.0000000
#> I9  0.0000000 0.7660520 0.0000000
#> I10 0.0000000 0.7026166 0.0000000
#> I11 0.0000000 0.7826122 0.1824270
#> I12 0.0000000 0.7957693 0.1877502
#> I13 0.0000000 0.0000000 0.7059281
#> I14 0.0000000 0.0000000 0.7378874
#> I15 0.0000000 0.0000000 0.7227175
#> I16 0.0000000 0.0000000 0.6789860
#> I17 0.0000000 0.0000000 0.6725952
#> I18 0.0000000 0.0000000 0.6753587

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

# }