Summary method for pcfa objectects

Provide basic information for an PCFA object, and summarize various posteriors.

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

shrink: (Ave) shrinkage for each facor's loadings and LD (if LD in pcfa = T).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

# \donttest{
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 500
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 20
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.108979   1.198945
#> F2   1.009585   1.015587
#> F3   2.869802   8.396947
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F       est         sd     lower     upper sig
#> var1     1 1 0.7132096 0.07559757 0.5826406 0.8762726   1
#> var2     2 1 0.7151372 0.07469996 0.5815766 0.8624938   1
#> var3     3 1 0.7206282 0.07661523 0.5728306 0.8625480   1
#> var4     4 1 0.6801365 0.07554690 0.5295291 0.8207183   1
#> var5     5 1 0.7516212 0.09558708 0.5841590 0.9370623   1
#> var6     6 1 0.7207744 0.09856962 0.5460564 0.9116566   1
#> var25    7 2 0.6753544 0.05634739 0.5746276 0.7921577   1
#> var26    8 2 0.7292697 0.05586815 0.6240301 0.8389137   1
#> var27    9 2 0.6914973 0.05400945 0.5838991 0.7968323   1
#> var28   10 2 0.6387546 0.05780460 0.5303563 0.7582456   1
#> var29   11 2 0.6676210 0.07437936 0.4998419 0.8069145   1
#> var30   12 2 0.6457860 0.08366174 0.4594421 0.8088702   1
#> var47   11 3 0.4448189 0.08413204 0.2643365 0.5833853   1
#> var48   12 3 0.4906045 0.08213580 0.3148425 0.6426463   1
#> var49   13 3 0.6776421 0.08250375 0.5265160 0.8391983   1
#> var50   14 3 0.7253924 0.08094882 0.5929502 0.8988512   1
#> var51   15 3 0.7549553 0.08722649 0.6048518 0.9257804   1
#> var52   16 3 0.8309306 0.09169774 0.6776833 1.0110218   1
#> var53   17 3 0.7994678 0.10974540 0.6062078 0.9692262   1
#> var54   18 3 0.7898072 0.11131827 0.5896706 0.9671319   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]      [,3]
#> I1  0.7132096 0.0000000 0.0000000
#> I2  0.7151372 0.0000000 0.0000000
#> I3  0.7206282 0.0000000 0.0000000
#> I4  0.6801365 0.0000000 0.0000000
#> I5  0.7516212 0.0000000 0.0000000
#> I6  0.7207744 0.0000000 0.0000000
#> I7  0.0000000 0.6753544 0.0000000
#> I8  0.0000000 0.7292697 0.0000000
#> I9  0.0000000 0.6914973 0.0000000
#> I10 0.0000000 0.6387546 0.0000000
#> I11 0.0000000 0.6676210 0.4448189
#> I12 0.0000000 0.6457860 0.4906045
#> I13 0.0000000 0.0000000 0.6776421
#> I14 0.0000000 0.0000000 0.7253924
#> I15 0.0000000 0.0000000 0.7549553
#> I16 0.0000000 0.0000000 0.8309306
#> I17 0.0000000 0.0000000 0.7994678
#> I18 0.0000000 0.0000000 0.7898072

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

summary(m0, what = 'all') #summarize all information

#> $N
#> [1] 500
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 20
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.108979   1.198945
#> F2   1.009585   1.015587
#> F3   2.869802   8.396947
#> 
#> $lambda
#>       Item F       est         sd     lower     upper sig
#> var1     1 1 0.7132096 0.07559757 0.5826406 0.8762726   1
#> var2     2 1 0.7151372 0.07469996 0.5815766 0.8624938   1
#> var3     3 1 0.7206282 0.07661523 0.5728306 0.8625480   1
#> var4     4 1 0.6801365 0.07554690 0.5295291 0.8207183   1
#> var5     5 1 0.7516212 0.09558708 0.5841590 0.9370623   1
#> var6     6 1 0.7207744 0.09856962 0.5460564 0.9116566   1
#> var25    7 2 0.6753544 0.05634739 0.5746276 0.7921577   1
#> var26    8 2 0.7292697 0.05586815 0.6240301 0.8389137   1
#> var27    9 2 0.6914973 0.05400945 0.5838991 0.7968323   1
#> var28   10 2 0.6387546 0.05780460 0.5303563 0.7582456   1
#> var29   11 2 0.6676210 0.07437936 0.4998419 0.8069145   1
#> var30   12 2 0.6457860 0.08366174 0.4594421 0.8088702   1
#> var47   11 3 0.4448189 0.08413204 0.2643365 0.5833853   1
#> var48   12 3 0.4906045 0.08213580 0.3148425 0.6426463   1
#> var49   13 3 0.6776421 0.08250375 0.5265160 0.8391983   1
#> var50   14 3 0.7253924 0.08094882 0.5929502 0.8988512   1
#> var51   15 3 0.7549553 0.08722649 0.6048518 0.9257804   1
#> var52   16 3 0.8309306 0.09169774 0.6776833 1.0110218   1
#> var53   17 3 0.7994678 0.10974540 0.6062078 0.9692262   1
#> var54   18 3 0.7898072 0.11131827 0.5896706 0.9671319   1
#> 
#> $qlambda
#>          [,1]      [,2]      [,3]
#> I1  0.7132096 0.0000000 0.0000000
#> I2  0.7151372 0.0000000 0.0000000
#> I3  0.7206282 0.0000000 0.0000000
#> I4  0.6801365 0.0000000 0.0000000
#> I5  0.7516212 0.0000000 0.0000000
#> I6  0.7207744 0.0000000 0.0000000
#> I7  0.0000000 0.6753544 0.0000000
#> I8  0.0000000 0.7292697 0.0000000
#> I9  0.0000000 0.6914973 0.0000000
#> I10 0.0000000 0.6387546 0.0000000
#> I11 0.0000000 0.6676210 0.4448189
#> I12 0.0000000 0.6457860 0.4906045
#> I13 0.0000000 0.0000000 0.6776421
#> I14 0.0000000 0.0000000 0.7253924
#> I15 0.0000000 0.0000000 0.7549553
#> I16 0.0000000 0.0000000 0.8309306
#> I17 0.0000000 0.0000000 0.7994678
#> I18 0.0000000 0.0000000 0.7898072
#> 
#> $eigen
#>         est        sd    lower    upper sig
#> F1 3.608641 0.6782755 2.430637 5.032449   1
#> F2 3.145830 0.4004690 2.460953 3.930004   1
#> F3 4.167623 0.9300883 2.801022 5.883478   1
#> 
#> $dpsx
#>           est         sd     lower     upper sig
#> I1  0.5147911 0.03472185 0.4432788 0.5782381   1
#> I2  0.5316776 0.03668399 0.4646589 0.6067739   1
#> I3  0.5081358 0.03667765 0.4350894 0.5777886   1
#> I4  0.5338122 0.03683461 0.4653966 0.6083858   1
#> I5  0.1763382 0.01779228 0.1414804 0.2105950   1
#> I6  0.1928909 0.01756231 0.1607099 0.2275649   1
#> I7  0.5458289 0.03897775 0.4789994 0.6273314   1
#> I8  0.4924882 0.03626421 0.4268947 0.5668412   1
#> I9  0.4876025 0.03526662 0.4215090 0.5581236   1
#> I10 0.5665360 0.04021554 0.4923415 0.6449224   1
#> I11 0.1815831 0.01695060 0.1477644 0.2148859   1
#> I12 0.1873897 0.01739314 0.1559123 0.2227636   1
#> I13 0.5965901 0.04017484 0.5273136 0.6852119   1
#> I14 0.5336932 0.03681814 0.4631180 0.6056358   1
#> I15 0.5165855 0.03667982 0.4413151 0.5842685   1
#> I16 0.4381103 0.03215554 0.3779437 0.4980445   1
#> I17 0.1832037 0.01691374 0.1507237 0.2160482   1
#> I18 0.2071104 0.01859753 0.1716268 0.2455281   1
#> 
#> $phi
#>      row col       est        sd       lower     upper sig
#> var1   2   1 0.3368614 0.2078506 -0.05618806 0.6529444   0
#> var2   3   1 0.4337619 0.1422856  0.18511999 0.7224248   1
#> var3   3   2 0.3495341 0.1137156  0.12513573 0.5770617   1
#> 
#> $shrink
#>        est        sd    lower    upper sig
#> F1  3.0207 0.5128796 2.083158 4.043662   1
#> F2  3.0207 0.5128796 2.083158 4.043662   1
#> F3  3.0207 0.5128796 2.083158 4.043662   1
#> PSX 0.0000 0.0000000 0.000000 0.000000   0
#> 

# }