Summary method for lawbl objects

Provide summaries of posterior information for a lawbl object, .

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

thd: Threshold estimates.

int: Intercept estimates (for pcirm only).

shrink: (Ave) shrinkage for each factor's loadings and LD (if LD in pcfa = T).

factor: Are the factors true of not (for EFA).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

# \donttest{
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[7:8,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 1000, iter = 1000)
summary(m0) # summarize basic information

#> $N
#> [1] 1000
#> 
#> $J
#> [1] 18
#> 
#> $K
#> [1] 3
#> 
#> $`Miss%`
#> [1] 0
#> 
#> $`LD enabled`
#> [1] FALSE
#> 
#> $`Burn in`
#> [1] 1000
#> 
#> $Iteration
#> [1] 1000
#> 
#> $`No. of sig lambda`
#> [1] 23
#> 
#> $`True Factor`
#> [1] TRUE TRUE TRUE
#> 
#> $`Adj. PSR`
#>    Point est. Upper C.I.
#> F1   1.392798   2.747404
#> F2   2.283015   6.579482
#> F3   1.801771   3.783331
#> 

summary(m0, what = 'lambda') #summarize significant loadings

#>       Item F        est         sd       lower        upper sig
#> var1     1 1  0.7168318 0.05863130  0.60063983  0.834773268   1
#> var2     2 1  0.7278316 0.05779678  0.64013468  0.871120556   1
#> var3     3 1  0.7141098 0.05715956  0.61875267  0.839220462   1
#> var4     4 1  0.6825221 0.05641237  0.57940630  0.803445197   1
#> var5     5 1  0.7209264 0.07175107  0.58625947  0.881376142   1
#> var6     6 1  0.7090946 0.07071346  0.57962200  0.876582492   1
#> var17   17 1  0.3988808 0.05916764  0.27487582  0.510634374   1
#> var18   18 1  0.3873534 0.05822375  0.26710952  0.501540140   1
#> var23    5 2  0.3944824 0.15283050  0.05486047  0.636611569   1
#> var24    6 2  0.3838388 0.14373765  0.05859535  0.600339306   1
#> var25    7 2  0.7794717 0.08995055  0.65421472  0.985734195   1
#> var26    8 2  0.7573627 0.09306010  0.61409623  0.971493566   1
#> var27    9 2  0.7838014 0.09591658  0.63355807  0.976700546   1
#> var28   10 2  0.7916724 0.10275099  0.61551683  0.993459553   1
#> var29   11 2  0.8078252 0.07340872  0.66921092  0.964513776   1
#> var30   12 2  0.8068142 0.07248427  0.67653709  0.968216406   1
#> var46   10 3 -0.2371395 0.14284757 -0.52744894 -0.001568496   1
#> var49   13 3  0.6985774 0.04795899  0.61212920  0.795994692   1
#> var50   14 3  0.6623036 0.04995888  0.56516558  0.761825928   1
#> var51   15 3  0.6895929 0.05092512  0.60313158  0.808199124   1
#> var52   16 3  0.6844029 0.05024377  0.58734395  0.783395943   1
#> var53   17 3  0.7137287 0.05610414  0.59888817  0.811465827   1
#> var54   18 3  0.7056370 0.05391711  0.60763129  0.812940899   1

summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format

#>          [,1]      [,2]       [,3]
#> I1  0.7168318 0.0000000  0.0000000
#> I2  0.7278316 0.0000000  0.0000000
#> I3  0.7141098 0.0000000  0.0000000
#> I4  0.6825221 0.0000000  0.0000000
#> I5  0.7209264 0.3944824  0.0000000
#> I6  0.7090946 0.3838388  0.0000000
#> I7  0.0000000 0.7794717  0.0000000
#> I8  0.0000000 0.7573627  0.0000000
#> I9  0.0000000 0.7838014  0.0000000
#> I10 0.0000000 0.7916724 -0.2371395
#> I11 0.0000000 0.8078252  0.0000000
#> I12 0.0000000 0.8068142  0.0000000
#> I13 0.0000000 0.0000000  0.6985774
#> I14 0.0000000 0.0000000  0.6623036
#> I15 0.0000000 0.0000000  0.6895929
#> I16 0.0000000 0.0000000  0.6844029
#> I17 0.3988808 0.0000000  0.7137287
#> I18 0.3873534 0.0000000  0.7056370

summary(m0, what = 'offpsx') #summarize significant LD terms

#> NULL

# }