Summary method for pcfa objectects

Provide basic information for an PCFA object, and summarize various posteriors.

# S3 method for lawbl
summary(
  object,
  what = "basic",
  med = FALSE,
  SL = 0.05,
  detail = FALSE,
  digits = 4,
  ...
)

Arguments

object
A lawbl object

what
A list of options for what to summarize.

basic: Basic information about the model and posteriors.

lambda: Loading estimates.

qlambda: Loading estimates in pattern/Q-matrix format.

eigen: Factorial eigen value.

dpsx: Diagonal elements in the residual covariance matrix PSX.

offpsx: Off-diagonal elements in PSX; local dependence terms.

phi: Factorial correlations.

shrink: (Ave) shrinkage for each facor's loadings and LD (if LD in pcfa = T).

all: All above information.

med
logical; if the posterior median (TRUE) or mean (FALSE) is used as the estimate.

SL
Significance level for interval estimate. The default is .05.

detail
logical; if only significant (FALSE) or all (TRUE) estimates are presented.

digits
Number of significant digits to print when printing numeric values.

...
additional arguments

Value

A list or matrix containing the summarized information based on the option what.

Examples

if (FALSE) {
dat <- sim18cfa0$dat
J <- ncol(dat)
K <- 3
Q<-matrix(-1,J,K);
Q[1:2,1]<-Q[9:10,2]<-Q[13:14,3]<-1

m0 <- pcfa(dat = dat, Q = Q, LD = FALSE,burn = 2000, iter = 2000)
summary(m0) # summarize basic information
summary(m0, what = 'lambda') #summarize significant loadings
summary(m0, what = 'qlambda') #summarize significant loadings in pattern/Q-matrix format
summary(m0, what = 'offpsx') #summarize significant LD terms
summary(m0, what = 'all') #summarize all information
}